Split crt drivers into separate files and satisfy lama
