+++ src/beeflow/common/worker/worker.py
@@ -10,11 +10,11 @@ from beeflow.common.crt_interface import ContainerRuntimeInterface
 # Import all implemented container runtime drivers now
 # No error if they don't exist
 try:
-    from beeflow.common.crt_drivers import CharliecloudDriver
+    from beeflow.common.crt.charliecloud_driver import CharliecloudDriver
 except ModuleNotFoundError:
     pass
 try:
-    from beeflow.common.crt_drivers import SingularityDriver
+    from beeflow.common.crt.singularity_driver import SingularityDriver
 except ModuleNotFoundError:
     pass
